Add postgres

+++ helper.py
@@ -1,9 +1,17 @@
+import os
+
+SQLITE3_ENABLED = True if os.environ['SQLITE3_ENABLED'] == True else False
+
+def prepareQuery(query):
+    if SQLITE3_ENABLED == False:
+        return query.replace('?', '%s')
+
 def getUser(conn, author):
     """
     Query the User
     """
     c = conn.cursor()
-    c.execute("SELECT user_id, COUNT(*) as votes from votes_history WHERE user_id=?", (int(author), ))
+    c.execute(prepareQuery("SELECT user_id, COUNT(*) as votes from votes_history WHERE user_id=?"), (int(author), ))
     row = c.fetchone()
     return row
 
@@ -13,7 +21,7 @@ def updateHistory(conn, author, message_id, backer):
     Returns success
     """
     c = conn.cursor()
-    c.execute("INSERT INTO votes_history (user_id, message_id, backer) VALUES (?,?,?)", (int(author), int(message_id), int(backer), ))
+    c.execute(prepareQuery("INSERT INTO votes_history (user_id, message_id, backer) VALUES (?,?,?)"), (int(author), int(message_id), int(backer), ))
     conn.commit()
     return c.rowcount > 0
 
@@ -24,9 +32,9 @@ def removeHistory(conn, message_id, author = None, backer = None):
     """
     c = conn.cursor()
     if backer != None and author != None:
-        c.execute("DELETE FROM votes_history WHERE user_id=? AND message_id=? AND backer=?", (int(author), int(message_id), int(backer), ))
+        c.execute(prepareQuery("DELETE FROM votes_history WHERE user_id=? AND message_id=? AND backer=?"), (int(author), int(message_id), int(backer), ))
     else:
-        c.execute("DELETE FROM votes_history WHERE message_id=?", (int(message_id), ))
+        c.execute(prepareQuery("DELETE FROM votes_history WHERE message_id=?"), (int(message_id), ))
     conn.commit()
     return c.rowcount > 0
 
@@ -35,7 +43,7 @@ def fetchLeaderboard(conn):
     Returns the current leaderboard
     """
     c = conn.cursor()
-    c.execute("SELECT user_id, COUNT() as votes FROM votes_history ORDER by votes DESC")
+    c.execute(prepareQuery("SELECT user_id, COUNT() as votes FROM votes_history ORDER by votes DESC"))
     return c.fetchall()
 
 def queryHistory(conn, message_id):
@@ -43,6 +51,6 @@ def queryHistory(conn, message_id):
     Query for a Specific Message
     """
     c = conn.cursor()
-    c.execute("SELECT user_id from votes_history WHERE message_id=?", (int(message_id), ))
+    c.execute(prepareQuery("SELECT user_id from votes_history WHERE message_id=?", (int(message_id), )))
     row = c.fetchone()
     return row
